Face Attendance/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Werkzeug==2.0.1
SQLAlchemy==1.4.23
face-recognition==1.3.0
opencv-python==4.5.3.56
numpy==1.21.2
Pillow==8.3.2 